--- kafka/oauth/__init__.py
@@ -1,3 +1,0 @@
-from __future__ import absolute_import
-
-from kafka.oauth.abstract import AbstractTokenProvider
